refactor: websocket method
